Revert "Entrega parcial, hexagonal, falta ajustar gerentes e contas e atualizar os testes;"
