+++ .gitignore
@@ -19,8 +19,6 @@ lerna-debug.log*
 /coverage
 /.nyc_output
 
-
-
 # IDEs and editors
 /.idea
 .project
@@ -43,10 +41,6 @@ lerna-debug.log*
 .env.test.local
 .env.production.local
 .env.local
-
-#BD
-/ormconfig.json
-
 
 # temp directory
 .temp
