--- src/Application/dtos/clientes/update-cliente.dto.ts
@@ -1,9 +1,0 @@
-import { TipoConta } from '../../../domain/enums/tipo-conta-enum';
-
-export class UpdateClienteDto {
-  nomeCompleto: string;
-  endereco?: string;
-  telefone?: string;
-  gerenteId?: string;
-  tipoConta?: TipoConta;
-}
