--- src/Application/dtos/contas/create-conta.dto.ts
@@ -1,9 +1,0 @@
-import { TipoConta } from 'src/domain/enums/tipo-conta-enum';
-
-export class CreateContaDto {
-  clienteId: string;
-  tipoConta: TipoConta;
-  saldoInicial: number;
-  limiteChequeEspecial?: number; 
-  taxaJuros?: number; 
-}
