--- src/Application/dtos/contas/update-conta.dto.ts
@@ -1,12 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateContaDto } from './create-conta.dto';
-import { TipoConta } from 'src/domain/enums/tipo-conta-enum';
-
-export class UpdateContaDto extends PartialType(CreateContaDto) {
-
-  clienteId: string;
-  tipoConta: TipoConta;
-  saldoInicial?: number;
-  limiteChequeEspecial?: number; 
-  taxaJuros?: number; 
-}
