--- src/Application/dtos/gerentes/create-gerente.dto.ts
@@ -1,7 +1,0 @@
-import { Cliente } from '../../../domain/clientes/entities/cliente.entity'; 
-
-export class CreateGerenteDto {
-    nomeCompleto: string;
-    id: string;
-    clientes: Cliente[];
-}
