--- src/Application/dtos/gerentes/update-gerente.dto.ts
@@ -1,7 +1,0 @@
-import { Cliente } from "src/domain/clientes/entities/cliente.entity";
-
-export class UpdateGerenteDto  {
-    nomeCompleto: string;
-    id: string;
-    clientes?: Cliente[];
-}
