--- src/domain/contas/entities/ContaCorrente.entity.ts
@@ -1,12 +1,0 @@
-import { ChildEntity, Column } from 'typeorm';
-import { ContaBancaria } from './Conta.entity';
-
-@ChildEntity()
-export class ContaCorrente extends ContaBancaria {
-  @Column()
-  limiteChequeEspecial: number;
-
-  @Column()
-  saldo: number;
-
-}
