+++ src/dominio/clientes/dto/update-cliente.dto.ts
@@ -1,0 +1,10 @@
+import { TipoConta } from '../../enums/tipo-conta-enum';
+
+export class UpdateClienteDto {
+  id: string;
+  nomeCompleto: string;
+  endereco?: string;
+  telefone?: string;
+  gerenteId?: string;
+  tipoConta?: TipoConta;
+}
