+++ src/dominio/contas/dto/create-conta.dto.ts
@@ -1,0 +1,9 @@
+import { TipoConta } from '../../enums/tipo-conta-enum';
+
+export class CreateContaDto {
+  clienteId: string;
+  tipoConta: TipoConta;
+  saldoInicial: number;
+  limiteChequeEspecial?: number; // Only for ContaCorrente
+  taxaJuros?: number; // Only for ContaPoupanca
+}
