+++ src/dominio/contas/dto/update-conta.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateContaDto } from './create-conta.dto';
+
+export class UpdateContaDto extends PartialType(CreateContaDto) {}
