+++ src/dominio/gerentes/dto/create-gerente.dto.ts
@@ -1,0 +1,7 @@
+import { Cliente } from '../../clientes/entities/cliente.entity'; 
+
+export class CreateGerenteDto {
+    nomeCompleto: string;
+    id: string;
+    clientes: Cliente[];
+}
