+++ src/dominio/gerentes/dto/update-gerente.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateGerenteDto } from './create-gerente.dto';
+
+export class UpdateGerenteDto extends PartialType(CreateGerenteDto) {}
