+++ src/infraestrutura/repositorios/clientes.repository.ts
@@ -1,0 +1,5 @@
+import { EntityRepository, Repository } from 'typeorm';
+import { Cliente } from '../../dominio/clientes/entities/cliente.entity';
+
+@EntityRepository(Cliente)
+export class ClienteRepository extends Repository<Cliente> {}
