+++ src/infraestrutura/repositorios/gerentes.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Gerente } from '../../domain/gerentes/entities/gerente.entity';
+import { Gerente } from '../../dominio/gerentes/entities/gerente.entity';
 
 @EntityRepository(Gerente)
 export class GerenteRepository extends Repository<Gerente> {}
