+++ src/tests/cliente.service.spec.ts
@@ -1,9 +1,9 @@
 import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppModule } from '../app.module';
+import { AppModule } from '../aplicação/app.module';
 import * as supertest from 'supertest';
-import { Cliente } from 'src/domain/clientes/entities/cliente.entity';
-import { TipoConta } from '../domain/enums/tipo-conta-enum';
+import { Cliente } from 'src/dominio/usuarios/cliente.module';
+import { TipoConta } from '../dominio/enums/tipo-conta-enum';
 
 describe('Criar cliente', () => {
   let app: INestApplication;
@@ -17,14 +17,11 @@ describe('Criar cliente', () => {
     await app.init();
   });
 
-  afterAll(async () => {
-    await app.close();
-  });
-
   test('Deve criar um cliente com sucesso e adicionar um tipo de conta poupança e associar um gerente ele', async () => {
     const nomeCompleto = 'Lais';
     const endereco = 'Rua paraibuna, 123';
     const telefone = '429858464';
+    const gerente = 'João';
     const tipoConta = TipoConta.POUPANCA;
 
     const response = await supertest(app.getHttpServer())
@@ -33,17 +30,18 @@ describe('Criar cliente', () => {
         nomeCompleto,
         endereco,
         telefone,
+        gerente,
         tipoConta,
       })
       .expect(201);
 
     const body = response.body;
     console.log(body);
-    expect(body).toHaveProperty('id');
     expect(body.nomeCompleto).toBe(nomeCompleto);
     expect(body.endereco).toBe(endereco);
     expect(body.telefone).toBe(telefone);
-    expect(body.contas[0].tipoConta).toBe(TipoConta.POUPANCA);
+    expect(body.gerente.nomeCompleto).toBe(gerente);
+    expect(body.contas[0].tipoConta).toBe(tipoConta);
   });
 
   test('Deve retornar uma lista de clientes', async () => {
@@ -51,27 +49,15 @@ describe('Criar cliente', () => {
       nomeCompleto: 'Teste Cliente 1',
       endereco: 'Endereço Teste 1',
       telefone: 'Telefone Teste 1',
-      gerente: {
-        nomeCompleto: 'Gerente Teste 1',
-      },
-      contas: [
-        {
-          tipoConta: TipoConta.CORRENTE,
-        },
-      ],
+      gerente: 'Gerente Teste 1',
+      tipoConta: TipoConta.CORRENTE,
     };
     const cliente2 = {
       nomeCompleto: 'Teste Cliente 2',
       endereco: 'Endereço Teste 2',
       telefone: 'Telefone Teste 2',
-      gerente: {
-        nomeCompleto: 'Gerente Teste 2',
-      },
-      contas: [
-        {
-          tipoConta: TipoConta.POUPANCA,
-        },
-      ],
+      gerente: 'Gerente Teste 2',
+      tipoConta: TipoConta.POUPANCA,
     };
     await supertest(app.getHttpServer())
       .post('/clientes')
@@ -94,7 +80,7 @@ describe('Criar cliente', () => {
       expect(cliente).toHaveProperty('telefone');
       expect(cliente).toHaveProperty('contas');
       expect(cliente.contas[0]).toHaveProperty('tipoConta');
-      expect(cliente.gerenteId).toHaveProperty('nomeCompleto');
+      expect(cliente.gerente).toHaveProperty('nomeCompleto');
     });
   });
 });
